Edit dan Delete

+++ ul-laravel/project_6-kasir/database/migrations/2023_11_13_034849_create_table_barang.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('kategori_id');
             $table->string('nama_barang');
-            $table->decimal('harga');
+            $table->bigInteger('harga');
             $table->integer('jumlah_barang');
             $table->string('gambar')->nullable();
             $table->timestamps();
